# backend/api-server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "2017-09-20T19:40:58Z",
    "title": "api.sequencediagram.io"
  },
  "host": "api.sequencediagram.io",
  "basePath": "/unstable",
  "schemes": ["https"],
  "paths": {
    "/sequencediagrams": {
      "post": {
        "operationId": "CreateSequenceDiagram",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "SequenceDiagram",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SequenceDiagram"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/SequenceDiagramResponse"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri":
            "arn:aws:apigateway:eu-west-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-1:996553720565:function:test/invocations",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST",
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Empty"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods":
                  "'POST,GET,OPTIONS'",
                "method.response.header.Access-Control-Allow-Headers":
                  "'Content-Type,X-Amz-Date,Authorization,X-Api-Key'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    },
    "/sequencediagrams/{sequenceDiagramId}": {
      "get": {
        "operationId": "GetSequenceDiagram",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "sequenceDiagramId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/SequenceDiagramResponse"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri":
            "arn:aws:apigateway:eu-west-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-1:996553720565:function:test/invocations",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST",
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "post": {
        "operationId": "CreateSequenceDiagramRevision",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "SequenceDiagram",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SequenceDiagram"
            }
          },
          {
            "name": "sequenceDiagramId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/SequenceDiagramResponse"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri":
            "arn:aws:apigateway:eu-west-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-1:996553720565:function:test/invocations",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST",
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "sequenceDiagramId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Empty"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods":
                  "'GET,OPTIONS'",
                "method.response.header.Access-Control-Allow-Headers":
                  "'Content-Type,X-Amz-Date,Authorization,X-Api-Key'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    },
    "/sequencediagrams/{sequenceDiagramId}/{revision}": {
      "get": {
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sequenceDiagramId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/SequenceDiagramResponse"
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri":
            "arn:aws:apigateway:eu-west-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-1:996553720565:function:test/invocations",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST",
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sequenceDiagramId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Empty"
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              },
              "Access-Control-Allow-Methods": {
                "type": "string"
              },
              "Access-Control-Allow-Headers": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods":
                  "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'",
                "method.response.header.Access-Control-Allow-Headers":
                  "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    }
  },
  "definitions": {
    "Empty": {
      "type": "object"
    },
    "ObjectId": {
      "type": "string",
      "pattern": "^o[0-9]+$"
    },
    "MessageId": {
      "type": "string",
      "pattern": "^m[0-9]+$"
    },
    "Object": {
      "type": "object",
      "required": ["id", "name"],
      "additionalProperties": false,
      "properties": {
        "id": {
          "$ref": "#/definitions/ObjectId"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Message": {
      "type": "object",
      "required": ["id", "sender", "receiver", "name"],
      "additionalProperties": false,
      "properties": {
        "id": {
          "$ref": "#/definitions/MessageId"
        },
        "name": {
          "type": "string"
        },
        "sender": {
          "$ref": "#/definitions/ObjectId"
        },
        "receiver": {
          "$ref": "#/definitions/ObjectId"
        },
        "isAsync": {
          "type": "boolean"
        },
        "isReply": {
          "type": "boolean"
        }
      }
    },
    "SequenceDiagram": {
      "type": "object",
      "required": ["messages", "objects"],
      "additionalProperties": false,
      "properties": {
        "objects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Object"
          }
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      }
    },
    "SequenceDiagramResponse": {
      "type": "object",
      "required": ["id", "revision", "sequenceDiagram", "creationTimeUtc"],
      "additionalProperties": false,
      "properties": {
        "id": {
          "type": "string"
        },
        "revision": {
          "type": "integer"
        },
        "sequenceDiagram": {
          "$ref": "#/definitions/SequenceDiagram"
        },
        "creationTimeUtc": {
          "type": "integer"
        }
      }
    }
  }
}